new file feature-flags_test and fix lint
